Ciweilization 1.07: The end turn button only works for the active player now and the Game Manger recognizes the number of players in the game and behaves accordingly.
(1) The previous commit actually has no bugs, it seemed like there was a bug dur to bad connections to the server.
(2) Also for the previous commit many hero power/wonder ability triggering times were added to prepare for adding the abilities in.
(3) Tested that some wonders scripts actually works now in the appropriate triggering times., so that was great!