Ciweilization 1.11
(1) Implemented The Opportunist, whose hero power cannot be manually used in current build. (Now it's automatically built in and so it works!)
(2) Added a new hero, The Patient Kitty, and implemented it.
(3) Added a upper limit for the number of moves in a turn, and set it as 6 for now.